third_party/blink/web_tests/virtual/document-isolation-policy/external/wpt/html/document-isolation-policy/service-worker-coep-none-proxy.https.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] COEP:unsafe-none ServiceWorker
  assert_equals: expected "Document: Fetch error" but got "Document: Fetch success"
Harness: the test ran to completion.
